requirements.txt.txt
discord.py==1.6.0
colorama
tqdm